other-ideas/sail-summit-registration-complete/sail-summit-backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.13"}
psycopg = {extras = ["binary"], version = "^3.2.9"}
pydantic-settings = "^2.9.1"
stripe = "^12.2.0"
python-multipart = "^0.0.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
